Refactor error handling and node properties
