+++ packages/@n8n/nodes-langchain/nodes/agents/Agent/agents/ToolsAgent/execute.ts
@@ -20,7 +20,6 @@ import {
 	getConnectedTools,
 } from '../../../../../utils/helpers';
 import { SYSTEM_MESSAGE } from './prompt';
-import { getErrorMessageByStatusCode } from './error-handling';
 
 function getOutputParserSchema(outputParser: BaseOutputParser): ZodObject<any, any, any, any> {
 	const parserType = outputParser.lc_namespace[outputParser.lc_namespace.length - 1];
@@ -182,20 +181,12 @@ export async function toolsAgentExecute(this: IExecuteFunctions): Promise<INodeE
 				),
 			});
 		} catch (error) {
-			const customErrorMessage: string | undefined = getErrorMessageByStatusCode(
-				error?.response?.status,
-			);
-
 			if (this.continueOnFail(error)) {
 				returnData.push({
-					json: { error: customErrorMessage ?? error.message },
+					json: { error: error.message },
 					pairedItem: { item: itemIndex },
 				});
 				continue;
-			}
-
-			if (customErrorMessage) {
-				throw new NodeOperationError(this.getNode(), customErrorMessage);
 			}
 
 			throw error;
