+++ packages/@n8n/nodes-langchain/nodes/llms/LmChatGoogleVertex/LmChatGoogleVertex.node.ts
@@ -6,13 +6,16 @@ import {
 	type INodeTypeDescription,
 	type SupplyData,
 	type ILoadOptionsFunctions,
+	type JsonObject,
+	NodeOperationError,
 } from 'n8n-workflow';
 import { ChatVertexAI } from '@langchain/google-vertexai';
-import type { HarmBlockThreshold, HarmCategory, SafetySetting } from '@google/generative-ai';
+import type { SafetySetting } from '@google/generative-ai';
 import { ProjectsClient } from '@google-cloud/resource-manager';
 import { getConnectionHintNoticeField } from '../../../utils/sharedFields';
 import { N8nLlmTracing } from '../N8nLlmTracing';
-import { harmCategories, harmThresholds } from './options';
+import { getErrorMessageByStatusCode } from './error-handling';
+import { additionalOptions } from '../gemini-common/additional-options';
 
 export class LmChatGoogleVertex implements INodeType {
 	description: INodeTypeDescription = {
@@ -83,89 +86,7 @@ export class LmChatGoogleVertex implements INodeType {
 					'The model which will generate the completion. <a href="https://cloud.google.com/vertex-ai/generative-ai/docs/learn/models">Learn more</a>.',
 				default: 'gemini-1.5-flash',
 			},
-			{
-				displayName: 'Options',
-				name: 'options',
-				placeholder: 'Add Option',
-				description: 'Additional options to add',
-				type: 'collection',
-				default: {},
-				options: [
-					{
-						displayName: 'Maximum Number of Tokens',
-						name: 'maxOutputTokens',
-						default: 2048,
-						description: 'The maximum number of tokens to generate in the completion',
-						type: 'number',
-					},
-					{
-						displayName: 'Sampling Temperature',
-						name: 'temperature',
-						default: 0.4,
-						typeOptions: { maxValue: 1, minValue: 0, numberPrecision: 1 },
-						description:
-							'Controls randomness: Lowering results in less random completions. As the temperature approaches zero, the model will become deterministic and repetitive.',
-						type: 'number',
-					},
-					{
-						displayName: 'Top K',
-						name: 'topK',
-						default: 32,
-						typeOptions: { maxValue: 40, minValue: -1, numberPrecision: 1 },
-						description:
-							'Used to remove "long tail" low probability responses. Defaults to -1, which disables it.',
-						type: 'number',
-					},
-					{
-						displayName: 'Top P',
-						name: 'topP',
-						default: 1,
-						typeOptions: { maxValue: 1, minValue: 0, numberPrecision: 1 },
-						description:
-							'Controls diversity via nucleus sampling: 0.5 means half of all likelihood-weighted options are considered. We generally recommend altering this or temperature but not both.',
-						type: 'number',
-					},
-
-					// Safety Settings
-					{
-						displayName: 'Safety Settings',
-						name: 'safetySettings',
-						type: 'fixedCollection',
-						typeOptions: { multipleValues: true },
-						default: {
-							values: {
-								category: harmCategories[0].name as HarmCategory,
-								threshold: harmThresholds[0].name as HarmBlockThreshold,
-							},
-						},
-						placeholder: 'Add Option',
-						options: [
-							{
-								name: 'values',
-								displayName: 'Values',
-								values: [
-									{
-										displayName: 'Safety Category',
-										name: 'category',
-										type: 'options',
-										description: 'The category of harmful content to block',
-										default: 'HARM_CATEGORY_UNSPECIFIED',
-										options: harmCategories,
-									},
-									{
-										displayName: 'Safety Threshold',
-										name: 'threshold',
-										type: 'options',
-										description: 'The threshold of harmful content to block',
-										default: 'HARM_BLOCK_THRESHOLD_UNSPECIFIED',
-										options: harmThresholds,
-									},
-								],
-							},
-						],
-					},
-				],
-			},
+			additionalOptions,
 		],
 	};
 
@@ -226,25 +147,53 @@ export class LmChatGoogleVertex implements INodeType {
 			null,
 		) as SafetySetting[];
 
-		const model = new ChatVertexAI({
-			authOptions: {
-				projectId,
-				credentials: {
-					client_email: credentials.email as string,
-					private_key: credentials.privateKey as string,
+		try {
+			const model = new ChatVertexAI({
+				authOptions: {
+					projectId,
+					credentials: {
+						client_email: credentials.email as string,
+						private_key: credentials.privateKey as string,
+					},
 				},
-			},
-			model: modelName,
-			topK: options.topK,
-			topP: options.topP,
-			temperature: options.temperature,
-			maxOutputTokens: options.maxOutputTokens,
-			safetySettings,
-			callbacks: [new N8nLlmTracing(this)],
-		});
+				model: modelName,
+				topK: options.topK,
+				topP: options.topP,
+				temperature: options.temperature,
+				maxOutputTokens: options.maxOutputTokens,
+				safetySettings,
+				callbacks: [new N8nLlmTracing(this)],
+				// Handle ChatVertexAI invocation errors to provide better error messages
+				onFailedAttempt: (error: any) => {
+					const customErrorMessage: string | undefined = getErrorMessageByStatusCode(
+						error?.response?.status,
+					);
 
-		return {
-			response: model,
-		};
+					if (customErrorMessage) {
+						throw new NodeOperationError(this.getNode(), customErrorMessage);
+					}
+
+					throw error;
+				},
+			});
+
+			return {
+				response: model,
+			};
+		} catch (e) {
+			// Catch model name validation error from Langchain (https://github.com/langchain-ai/langchainjs/blob/ef201d0ee85ee4049078270a0cfd7a1767e624f8/libs/langchain-google-common/src/utils/common.ts#L124)
+			// to show more helpful error message
+			if (e?.message?.startsWith('Unable to verify model params')) {
+				throw new NodeOperationError(this.getNode(), e as JsonObject, {
+					message: 'Only models starting with "gemini" are supported',
+				});
+			}
+
+			// Assume all other exceptions while creating a new ChatVertexAI instance are parameter validation errors
+			throw new NodeOperationError(this.getNode(), e as JsonObject, {
+				message: 'Invalid options',
+				description: e.message,
+			});
+		}
 	}
 }
