+++ packages/@n8n/nodes-langchain/nodes/llms/gemini-common/safety-options.ts
