+++ packages/@n8n/nodes-langchain/package.json
@@ -138,7 +138,7 @@
     "@getzep/zep-cloud": "1.0.6",
     "@getzep/zep-js": "0.9.0",
     "@google-ai/generativelanguage": "2.5.0",
-    "@google-cloud/resource-manager": "^5.3.0",
+    "@google-cloud/resource-manager": "5.3.0",
     "@google/generative-ai": "0.11.4",
     "@huggingface/inference": "2.7.0",
     "@langchain/anthropic": "0.1.21",
@@ -146,7 +146,7 @@
     "@langchain/community": "0.2.13",
     "@langchain/core": "0.2.9",
     "@langchain/google-genai": "0.0.16",
-    "@langchain/google-vertexai": "^0.0.19",
+    "@langchain/google-vertexai": "0.0.19",
     "@langchain/groq": "0.0.12",
     "@langchain/mistralai": "0.0.22",
     "@langchain/openai": "0.0.33",
